+++ example/lib/crop_image_demo.dart
@@ -26,7 +26,6 @@ class _CropImageDemoState extends State<CropImageDemo> {
   @override
   void dispose() {
     listSourceRepository.dispose();
-    // TODO: implement dispose
     super.dispose();
   }