+++ example/lib/main.dart
@@ -62,7 +62,6 @@ class _MyHomePageState extends State<MyHomePage> {
   TuChongRepository listSourceRepository;
   @override
   void initState() {
-    // TODO: implement initState
     pages.add(Page(
         PageType.Image,
         "cache image"
@@ -89,7 +88,6 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
-    // TODO: implement build
     var content = ListView.builder(
       itemBuilder: (_, int index) {
         var page = pages[index];