+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Features:
   Add call back [onSlidingPage] when is sliding page, you can change other widgets state in page.[ExtendedImageSlidePage]
+* Add [enableSlideOutPage] parameter to define whether enable slide out page. [ExtendedImage]  
 
 ## [0.4.3]